requirements.txt
elasticsearch~=8.15.0
the-new-hotness~=1.3.0
mutagen~=1.47.0
pymongo~=4.14.1
python-dotenv~=1.1.1
kafka-python~=2.2.15
SpeechRecognition~=3.14.3